[CSC] Accept -1 from server as valid metrics
This CL updates the postmessage API for metrics to pass signed values
instead of unsigned. This is becasue the server sends -1 for unknown
values and client converts them to unsigned and clamps them to 10
which is a bug. This CL fixes that by changing all the params sent
to signed.

No server side change is required.

(cherry picked from commit 74730f15242cfc44ddee9198512c1ed4a9ec667c)

Bug: b/290684610
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4736573
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1179377}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4752627
Cr-Commit-Position: refs/branch-heads/5845@{#1152}
Cr-Branched-From: 5a5dff63a4a4c63b9b18589819bebb2566c85443-refs/heads/main@{#1160321}

+++ chrome/browser/companion/core/companion_metrics_logger.h
@@ -48,22 +48,22 @@ struct UiSurfaceMetrics {
   UiEvent last_event = UiEvent::kNotAvailable;
 
   // The position of the UI surface relative to the companion page.
-  size_t ui_surface_position = kInvalidPosition;
+  int ui_surface_position = kInvalidPosition;
 
   // The number of child elements that were considered to be shown within the
   // surface, e.g. number of candidate queries inside related queries component.
-  size_t child_element_available_count = kInvalidNumChildren;
+  int child_element_available_count = kInvalidNumChildren;
 
   // The number of child elements shown within the surface, e.g. number of
   // related queries inside related queries component.
-  size_t child_element_shown_count = kInvalidNumChildren;
+  int child_element_shown_count = kInvalidNumChildren;
 
   // The number of times user clicked on the surface.
-  uint32_t click_count = 0;
+  int click_count = 0;
 
   // The position of the clicked UI element within its parent list. Applicable
   // to surfaces that show a list.
-  int32_t click_position = kInvalidPosition;
+  int click_position = kInvalidPosition;
 };
 
 // Tracks visual suggestion metrics for a single page.
@@ -130,9 +130,9 @@ class CompanionMetricsLogger {
 
   // Logging method corresponding to `RecordUiSurfaceShown` in companion.mojom.
   void RecordUiSurfaceShown(UiSurface ui_surface,
-                            uint32_t ui_surface_position,
-                            uint32_t child_element_available_count,
-                            uint32_t child_element_shown_count);
+                            int32_t ui_surface_position,
+                            int32_t child_element_available_count,
+                            int32_t child_element_shown_count);
 
   // Logging method corresponding to `RecordUiSurfaceClicked` in
   // companion.mojom.
